RESOLVE p67257

+++ programmers/p67257.java
@@ -2,7 +2,7 @@ package programmers;
 
 import java.util.*;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public class p67257 {
     public static void main(String[] args) {
@@ -14,8 +14,8 @@ public class p67257 {
     public static long solution(String expression) {
         long answer = 0;
 
-        int[] numbers = Arrays.stream(expression.split("[-+*]"))
-                .mapToInt(Integer::parseInt)
+        long[] numbers = Arrays.stream(expression.split("[-+*]"))
+                .mapToLong(Long::parseLong)
                 .toArray();
         String[] operators = Arrays.stream(expression.split("\\d+"))
                 .filter(x -> !x.equals(""))
@@ -43,8 +43,8 @@ public class p67257 {
         };
     }
 
-    public static long calcByPriority(int[] numbers, String[] operators, String[] priority) {
-        List<Integer> numList = IntStream.of(numbers).boxed().collect(Collectors.toCollection(LinkedList::new));
+    public static long calcByPriority(long[] numbers, String[] operators, String[] priority) {
+        List<Long> numList = LongStream.of(numbers).boxed().collect(Collectors.toCollection(LinkedList::new));
         List<String> opList = Arrays.stream(operators).collect(Collectors.toCollection(LinkedList::new));
 
         for (String prior : priority) {
@@ -54,19 +54,19 @@ public class p67257 {
         return numList.get(0);
     }
 
-    public static void calcOnOperator(List<Integer> numList, List<String> opList, String curOperator) {
+    public static void calcOnOperator(List<Long> numList, List<String> opList, String curOperator) {
         for (int j=0; j<opList.size(); j++) {
             if (opList.get(j).equals(curOperator)) {
                 String operator = opList.remove(j);
-                int prevNum = numList.remove(j);
-                int nextNum = numList.remove(j);
+                long prevNum = numList.remove(j);
+                long nextNum = numList.remove(j);
                 numList.add(j, threeCalc(operator, prevNum, nextNum));
                 j--;
             }
         }
     }
 
-    public static int threeCalc(String operator, int prevNum, int nextNum) {
+    public static long threeCalc(String operator, long prevNum, long nextNum) {
         if (operator.equals("+")) return prevNum + nextNum;
         if (operator.equals("*")) return prevNum * nextNum;
         else return prevNum - nextNum;
